Fix quantification (implement De Bruijn indexes)
  Quantifying in the presence of nested quantifiers was broken.  Now,
when the application of a substitution goes under a quantifier, it
increments the internal indices of all generic variables in the
substitution to be greater than the number of quantified variables just
passed.  This way, when the types are instantiated, there is no
confusion about which generic variables are to be instantiated.